@@ -15,7 +15,7 @@ class _ChannelOnlinePaymentViewState extends State<ChannelOnlinePaymentView> {
   bool isChooseBanking = false;
 
   int changePageIndex = 0;
-  String paymentMethod;
+  late String paymentMethod;
 
   @override
   void initState() {
@@ -47,361 +47,364 @@ class _ChannelOnlinePaymentViewState extends State<ChannelOnlinePaymentView> {
         : Scaffold(
             backgroundColor: Color(0xff188FC5),
             body: SingleChildScrollView(
-              child: Column(
-                children: [
-                  SizedBox(
-                    height: 30,
-                  ),
-                  Text(
-                    'Choose your payment channels',
-                    style: TextStyle(
-                        fontWeight: FontWeight.bold,
-                        fontSize: 14,
-                        color: Color(0xffe9e9e9)),
-                  ),
-                  SizedBox(
-                    height: 30,
-                  ),
-                  Container(
-                      margin: EdgeInsets.only(left: 20, right: 20),
-                      padding: EdgeInsets.only(top: 5, bottom: 5),
-                      decoration: BoxDecoration(
-                        color: Color(0xfffafafa),
-                        borderRadius: BorderRadius.all(Radius.circular(5.0)),
-                      ),
-                      child: Column(
-                        children: [
-                          Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Row(
-                                children: [
-                                  SizedBox(
-                                    width: 20,
-                                  ),
-                                  Icon(
-                                    Icons.credit_card,
-                                    color: Color(0xff188FC5),
-                                    size: 40,
-                                  ),
-                                  SizedBox(
-                                    width: 10,
-                                  ),
-                                  Text(
-                                    'Cards',
-                                    style: TextStyle(
-                                        fontWeight: FontWeight.bold,
-                                        fontSize: 14,
-                                        color: Color(0xff188FC5)),
-                                  ),
-                                ],
-                              ),
-                              Padding(
-                                  padding: const EdgeInsets.only(right: 20),
-                                  child: !isChooseCard
-                                      ? GestureDetector(
-                                          onTap: () {
-                                            setState(() {
-                                              isChooseCard = true;
-                                              isChooseWallet = false;
-                                              isChooseCounter = false;
-                                              isChooseBanking = false;
-                                            });
-                                          },
-                                          child: Icon(
-                                            Icons.radio_button_off,
-                                            size: 15,
-                                            color: Color(0xff188FC5),
-                                          ),
-                                        )
-                                      : Icon(
-                                          Icons.radio_button_checked,
-                                          size: 15,
-                                          color: Color(0xff188FC5),
-                                        )),
-                            ],
-                          ),
-                          isChooseCard
-                              ? SizedBox(
-                                  height: 15,
-                                )
-                              : Container(),
-                          isChooseCard
-                              ? Image(
-                                  image:
-                                      AssetImage('assets/images/card_icon.png'))
-                              : Container(),
-                          isChooseCard
-                              ? SizedBox(
-                                  height: 15,
-                                )
-                              : Container(),
-                        ],
-                      )),
-                  SizedBox(
-                    height: 20,
-                  ),
-                  Container(
-                      margin: EdgeInsets.only(left: 20, right: 20),
-                      padding: EdgeInsets.only(top: 5, bottom: 5),
-                      decoration: BoxDecoration(
-                        color: Color(0xfffafafa),
-                        borderRadius: BorderRadius.all(Radius.circular(5.0)),
-                      ),
-                      child: Column(
-                        children: [
-                          Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Row(
-                                children: [
-                                  SizedBox(
-                                    width: 20,
-                                  ),
-                                  Icon(
-                                    Icons.credit_card,
-                                    color: Color(0xff188FC5),
-                                    size: 40,
-                                  ),
-                                  SizedBox(
-                                    width: 10,
-                                  ),
-                                  Text(
-                                    'Mobile Wallet',
-                                    style: TextStyle(
-                                        fontWeight: FontWeight.bold,
-                                        fontSize: 14,
-                                        color: Color(0xff188FC5)),
-                                  ),
-                                ],
-                              ),
-                              Padding(
-                                  padding: const EdgeInsets.only(right: 20),
-                                  child: !isChooseWallet
-                                      ? GestureDetector(
-                                          onTap: () {
-                                            setState(() {
-                                              isChooseCard = false;
-                                              isChooseWallet = true;
-                                              isChooseCounter = false;
-                                              isChooseBanking = false;
-                                            });
-                                          },
-                                          child: Icon(
-                                            Icons.radio_button_off,
-                                            size: 15,
-                                            color: Color(0xff188FC5),
-                                          ),
-                                        )
-                                      : Icon(
-                                          Icons.radio_button_checked,
-                                          size: 15,
-                                          color: Color(0xff188FC5),
-                                        )),
-                            ],
-                          ),
-                          isChooseWallet
-                              ? SizedBox(
-                                  height: 15,
-                                )
-                              : Container(),
-                          isChooseWallet
-                              ? Image(
-                                  image: AssetImage(
-                                      'assets/images/wallet_icon.png'))
-                              : Container(),
-                          isChooseWallet
-                              ? SizedBox(
-                                  height: 20,
-                                )
-                              : Container(),
-                        ],
-                      )),
-                  SizedBox(
-                    height: 20,
-                  ),
-                  Container(
-                      margin: EdgeInsets.only(left: 20, right: 20),
-                      padding: EdgeInsets.only(top: 5, bottom: 5),
-                      decoration: BoxDecoration(
-                        color: Color(0xfffafafa),
-                        borderRadius: BorderRadius.all(Radius.circular(5.0)),
-                      ),
-                      child: Column(
-                        children: [
-                          Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Row(
-                                children: [
-                                  SizedBox(
-                                    width: 20,
-                                  ),
-                                  Icon(
-                                    Icons.credit_card,
-                                    color: Color(0xff188FC5),
-                                    size: 40,
-                                  ),
-                                  SizedBox(
-                                    width: 10,
-                                  ),
-                                  Text(
-                                    'Counter Payment',
-                                    style: TextStyle(
-                                        fontWeight: FontWeight.bold,
-                                        fontSize: 14,
-                                        color: Color(0xff188FC5)),
-                                  ),
-                                ],
-                              ),
-                              Padding(
-                                  padding: const EdgeInsets.only(right: 20),
-                                  child: !isChooseCounter
-                                      ? GestureDetector(
-                                          onTap: () {
-                                            setState(() {
-                                              isChooseCard = false;
-                                              isChooseWallet = false;
-                                              isChooseCounter = true;
-                                              isChooseBanking = false;
-                                            });
-                                          },
-                                          child: Icon(
-                                            Icons.radio_button_off,
-                                            size: 15,
-                                            color: Color(0xff188FC5),
-                                          ),
-                                        )
-                                      : Icon(
-                                          Icons.radio_button_checked,
-                                          size: 15,
-                                          color: Color(0xff188FC5),
-                                        )),
-                            ],
-                          ),
-                          isChooseCounter
-                              ? SizedBox(
-                                  height: 15,
-                                )
-                              : Container(),
-                          isChooseCounter
-                              ? Image(
-                                  image: AssetImage(
-                                      'assets/images/counter_icon.png'))
-                              : Container(),
-                          isChooseCounter
-                              ? SizedBox(
-                                  height: 20,
-                                )
-                              : Container(),
-                        ],
-                      )),
-                  SizedBox(
-                    height: 20,
-                  ),
-                  Container(
-                      margin: EdgeInsets.only(left: 20, right: 20),
-                      padding: EdgeInsets.only(top: 5, bottom: 5),
-                      decoration: BoxDecoration(
-                        color: Color(0xfffafafa),
-                        borderRadius: BorderRadius.all(Radius.circular(5.0)),
-                      ),
-                      child: Column(
-                        children: [
-                          Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Row(
-                                children: [
-                                  SizedBox(
-                                    width: 20,
-                                  ),
-                                  Icon(
-                                    Icons.credit_card,
-                                    color: Color(0xff188FC5),
-                                    size: 40,
-                                  ),
-                                  SizedBox(
-                                    width: 10,
-                                  ),
-                                  Text(
-                                    'Mobile Banking/Internet Banking',
-                                    style: TextStyle(
-                                        fontWeight: FontWeight.bold,
-                                        fontSize: 14,
-                                        color: Color(0xff188FC5)),
-                                  ),
-                                ],
-                              ),
-                              Padding(
-                                  padding: const EdgeInsets.only(right: 20),
-                                  child: !isChooseBanking
-                                      ? GestureDetector(
-                                          onTap: () {
-                                            setState(() {
-                                              isChooseCard = false;
-                                              isChooseWallet = false;
-                                              isChooseCounter = false;
-                                              isChooseBanking = true;
-                                            });
-                                          },
-                                          child: Icon(
-                                            Icons.radio_button_off,
-                                            size: 15,
-                                            color: Color(0xff188FC5),
-                                          ),
-                                        )
-                                      : Icon(
-                                          Icons.radio_button_checked,
-                                          size: 15,
-                                          color: Color(0xff188FC5),
-                                        )),
-                            ],
-                          ),
-                          isChooseBanking
-                              ? SizedBox(
-                                  height: 15,
-                                )
-                              : Container(),
-                          isChooseBanking
-                              ? Image(
-                                  image: AssetImage(
-                                      'assets/images/banking_icon.png'))
-                              : Container(),
-                          isChooseBanking
-                              ? SizedBox(
-                                  height: 20,
-                                )
-                              : Container(),
-                        ],
-                      )),
-                  SizedBox(
-                    height: 60,
-                  ),
-                  Container(
-                    child: RaisedButton(
-                      onPressed: () {
-                        if(isChooseCard){
-                          AppUtils.showSuccessSnackBar('Success', 'Go to 2c2p Page');
-                        }
-                        else{
-                          setState(() {
-                            changePageIndex = 1;
-                          });
-                        }
-
-                      },
-                      child: Text('Confirm'),
-                      textColor: Color(0xffe9e9e9),
-                      color: isChooseCard ||
-                              isChooseWallet ||
-                              isChooseCounter ||
-                              isChooseBanking
-                          ? Color(0xffff5f1f)
-                          : Color(0xffaaaaaa),
+              child: Container(
+                margin: EdgeInsets.only(bottom: 40),
+                child: Column(
+                  children: [
+                    SizedBox(
+                      height: 30,
                     ),
-                  ),
-                ],
+                    Text(
+                      'Choose your payment channels',
+                      style: TextStyle(
+                          fontWeight: FontWeight.bold,
+                          fontSize: 14,
+                          color: Color(0xffe9e9e9)),
+                    ),
+                    SizedBox(
+                      height: 30,
+                    ),
+                    Container(
+                        margin: EdgeInsets.only(left: 20, right: 20),
+                        padding: EdgeInsets.only(top: 5, bottom: 5),
+                        decoration: BoxDecoration(
+                          color: Color(0xfffafafa),
+                          borderRadius: BorderRadius.all(Radius.circular(5.0)),
+                        ),
+                        child: Column(
+                          children: [
+                            Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [
+                                Row(
+                                  children: [
+                                    SizedBox(
+                                      width: 20,
+                                    ),
+                                    Icon(
+                                      Icons.credit_card,
+                                      color: Color(0xff188FC5),
+                                      size: 40,
+                                    ),
+                                    SizedBox(
+                                      width: 10,
+                                    ),
+                                    Text(
+                                      'Cards',
+                                      style: TextStyle(
+                                          fontWeight: FontWeight.bold,
+                                          fontSize: 14,
+                                          color: Color(0xff188FC5)),
+                                    ),
+                                  ],
+                                ),
+                                Padding(
+                                    padding: const EdgeInsets.only(right: 20),
+                                    child: !isChooseCard
+                                        ? GestureDetector(
+                                            onTap: () {
+                                              setState(() {
+                                                isChooseCard = true;
+                                                isChooseWallet = false;
+                                                isChooseCounter = false;
+                                                isChooseBanking = false;
+                                              });
+                                            },
+                                            child: Icon(
+                                              Icons.radio_button_off,
+                                              size: 15,
+                                              color: Color(0xff188FC5),
+                                            ),
+                                          )
+                                        : Icon(
+                                            Icons.radio_button_checked,
+                                            size: 15,
+                                            color: Color(0xff188FC5),
+                                          )),
+                              ],
+                            ),
+                            isChooseCard
+                                ? SizedBox(
+                                    height: 15,
+                                  )
+                                : Container(),
+                            isChooseCard
+                                ? Image(
+                                    image:
+                                        AssetImage('assets/images/card_icon.png'))
+                                : Container(),
+                            isChooseCard
+                                ? SizedBox(
+                                    height: 15,
+                                  )
+                                : Container(),
+                          ],
+                        )),
+                    SizedBox(
+                      height: 20,
+                    ),
+                    Container(
+                        margin: EdgeInsets.only(left: 20, right: 20),
+                        padding: EdgeInsets.only(top: 5, bottom: 5),
+                        decoration: BoxDecoration(
+                          color: Color(0xfffafafa),
+                          borderRadius: BorderRadius.all(Radius.circular(5.0)),
+                        ),
+                        child: Column(
+                          children: [
+                            Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [
+                                Row(
+                                  children: [
+                                    SizedBox(
+                                      width: 20,
+                                    ),
+                                    Icon(
+                                      Icons.credit_card,
+                                      color: Color(0xff188FC5),
+                                      size: 40,
+                                    ),
+                                    SizedBox(
+                                      width: 10,
+                                    ),
+                                    Text(
+                                      'Mobile Wallet',
+                                      style: TextStyle(
+                                          fontWeight: FontWeight.bold,
+                                          fontSize: 14,
+                                          color: Color(0xff188FC5)),
+                                    ),
+                                  ],
+                                ),
+                                Padding(
+                                    padding: const EdgeInsets.only(right: 20),
+                                    child: !isChooseWallet
+                                        ? GestureDetector(
+                                            onTap: () {
+                                              setState(() {
+                                                isChooseCard = false;
+                                                isChooseWallet = true;
+                                                isChooseCounter = false;
+                                                isChooseBanking = false;
+                                              });
+                                            },
+                                            child: Icon(
+                                              Icons.radio_button_off,
+                                              size: 15,
+                                              color: Color(0xff188FC5),
+                                            ),
+                                          )
+                                        : Icon(
+                                            Icons.radio_button_checked,
+                                            size: 15,
+                                            color: Color(0xff188FC5),
+                                          )),
+                              ],
+                            ),
+                            isChooseWallet
+                                ? SizedBox(
+                                    height: 15,
+                                  )
+                                : Container(),
+                            isChooseWallet
+                                ? Image(
+                                    image: AssetImage(
+                                        'assets/images/wallet_icon.png'))
+                                : Container(),
+                            isChooseWallet
+                                ? SizedBox(
+                                    height: 20,
+                                  )
+                                : Container(),
+                          ],
+                        )),
+                    SizedBox(
+                      height: 20,
+                    ),
+                    Container(
+                        margin: EdgeInsets.only(left: 20, right: 20),
+                        padding: EdgeInsets.only(top: 5, bottom: 5),
+                        decoration: BoxDecoration(
+                          color: Color(0xfffafafa),
+                          borderRadius: BorderRadius.all(Radius.circular(5.0)),
+                        ),
+                        child: Column(
+                          children: [
+                            Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [
+                                Row(
+                                  children: [
+                                    SizedBox(
+                                      width: 20,
+                                    ),
+                                    Icon(
+                                      Icons.credit_card,
+                                      color: Color(0xff188FC5),
+                                      size: 40,
+                                    ),
+                                    SizedBox(
+                                      width: 10,
+                                    ),
+                                    Text(
+                                      'Counter Payment',
+                                      style: TextStyle(
+                                          fontWeight: FontWeight.bold,
+                                          fontSize: 14,
+                                          color: Color(0xff188FC5)),
+                                    ),
+                                  ],
+                                ),
+                                Padding(
+                                    padding: const EdgeInsets.only(right: 20),
+                                    child: !isChooseCounter
+                                        ? GestureDetector(
+                                            onTap: () {
+                                              setState(() {
+                                                isChooseCard = false;
+                                                isChooseWallet = false;
+                                                isChooseCounter = true;
+                                                isChooseBanking = false;
+                                              });
+                                            },
+                                            child: Icon(
+                                              Icons.radio_button_off,
+                                              size: 15,
+                                              color: Color(0xff188FC5),
+                                            ),
+                                          )
+                                        : Icon(
+                                            Icons.radio_button_checked,
+                                            size: 15,
+                                            color: Color(0xff188FC5),
+                                          )),
+                              ],
+                            ),
+                            isChooseCounter
+                                ? SizedBox(
+                                    height: 15,
+                                  )
+                                : Container(),
+                            isChooseCounter
+                                ? Image(
+                                    image: AssetImage(
+                                        'assets/images/counter_icon.png'))
+                                : Container(),
+                            isChooseCounter
+                                ? SizedBox(
+                                    height: 20,
+                                  )
+                                : Container(),
+                          ],
+                        )),
+                    SizedBox(
+                      height: 20,
+                    ),
+                    Container(
+                        margin: EdgeInsets.only(left: 20, right: 20),
+                        padding: EdgeInsets.only(top: 5, bottom: 5),
+                        decoration: BoxDecoration(
+                          color: Color(0xfffafafa),
+                          borderRadius: BorderRadius.all(Radius.circular(5.0)),
+                        ),
+                        child: Column(
+                          children: [
+                            Row(
+                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                              children: [
+                                Row(
+                                  children: [
+                                    SizedBox(
+                                      width: 20,
+                                    ),
+                                    Icon(
+                                      Icons.credit_card,
+                                      color: Color(0xff188FC5),
+                                      size: 40,
+                                    ),
+                                    SizedBox(
+                                      width: 10,
+                                    ),
+                                    Text(
+                                      'Mobile Banking/Internet Banking',
+                                      style: TextStyle(
+                                          fontWeight: FontWeight.bold,
+                                          fontSize: 14,
+                                          color: Color(0xff188FC5)),
+                                    ),
+                                  ],
+                                ),
+                                Padding(
+                                    padding: const EdgeInsets.only(right: 20),
+                                    child: !isChooseBanking
+                                        ? GestureDetector(
+                                            onTap: () {
+                                              setState(() {
+                                                isChooseCard = false;
+                                                isChooseWallet = false;
+                                                isChooseCounter = false;
+                                                isChooseBanking = true;
+                                              });
+                                            },
+                                            child: Icon(
+                                              Icons.radio_button_off,
+                                              size: 15,
+                                              color: Color(0xff188FC5),
+                                            ),
+                                          )
+                                        : Icon(
+                                            Icons.radio_button_checked,
+                                            size: 15,
+                                            color: Color(0xff188FC5),
+                                          )),
+                              ],
+                            ),
+                            isChooseBanking
+                                ? SizedBox(
+                                    height: 15,
+                                  )
+                                : Container(),
+                            isChooseBanking
+                                ? Image(
+                                    image: AssetImage(
+                                        'assets/images/banking_icon.png'))
+                                : Container(),
+                            isChooseBanking
+                                ? SizedBox(
+                                    height: 20,
+                                  )
+                                : Container(),
+                          ],
+                        )),
+                    SizedBox(
+                      height: 60,
+                    ),
+                    Container(
+                      child: RaisedButton(
+                        onPressed: () {
+                          if(isChooseCard){
+                            AppUtils.showSuccessSnackBar('Success', 'Go to 2c2p Page');
+                          }
+                          else{
+                            setState(() {
+                              changePageIndex = 1;
+                            });
+                          }
+
+                        },
+                        child: Text('Confirm'),
+                        textColor: Color(0xffe9e9e9),
+                        color: isChooseCard ||
+                                isChooseWallet ||
+                                isChooseCounter ||
+                                isChooseBanking
+                            ? Color(0xffff5f1f)
+                            : Color(0xffaaaaaa),
+                      ),
+                    ),
+                  ],
+                ),
               ),
             ));
   }
