@@ -3,7 +3,7 @@ import 'package:MJN/models/invoiceVO.dart';
 import 'package:get/state_manager.dart';
 
 class InvoiceController extends GetxController{
-  InvoiceVo invoiceVo;
+  late InvoiceVo invoiceVo;
   var isLoading = true.obs;
 
   void fetchInvoiceDataByID(String token,String uid,String invoiceID) async {
