@@ -6,12 +6,12 @@ String invoiceVoToJson(InvoiceVo data) => json.encode(data.toJson());
 
 class InvoiceVo {
   InvoiceVo({
-    this.status,
-    this.responseCode,
-    this.description,
-    this.isRequieredUpdate,
-    this.isforceUpdate,
-    this.details,
+    required this.status,
+    required  this.responseCode,
+    required this.description,
+    required this.isRequieredUpdate,
+    required this.isforceUpdate,
+    required this.details,
   });
 
   String status;
@@ -42,13 +42,13 @@ class InvoiceVo {
 
 class Details {
   Details({
-    this.invoiceId,
-    this.paymentTotal,
-    this.currencyType,
-    this.startDate,
-    this.endDate,
-    this.creationDate,
-    this.modifiedDate,
+    required   this.invoiceId,
+    required  this.paymentTotal,
+    required this.currencyType,
+    required  this.startDate,
+    required this.endDate,
+    required  this.creationDate,
+    required  this.modifiedDate,
   });
 
   String invoiceId;
