@@ -6,12 +6,12 @@ String transactionListVoToJson(TransactionListVo data) => json.encode(data.toJso
 
 class TransactionListVo {
   TransactionListVo({
-    this.status,
-    this.responseCode,
-    this.description,
-    this.isRequieredUpdate,
-    this.isforceUpdate,
-    this.details,
+    required  this.status,
+    required this.responseCode,
+    required this.description,
+    required this.isRequieredUpdate,
+    required this.isforceUpdate,
+    required this.details,
   });
 
   String status;
@@ -42,15 +42,15 @@ class TransactionListVo {
 
 class Detail {
   Detail({
-    this.transactionId,
-    this.paymentTotal,
-    this.currencyType,
-    this.startDate,
-    this.endDate,
-    this.paidDate,
-    this.creationDate,
-    this.modifiedDate,
-    this.isCreditNote,
+    required this.transactionId,
+    required this.paymentTotal,
+    required this.currencyType,
+    required this.startDate,
+    required this.endDate,
+    required this.paidDate,
+    required this.creationDate,
+    required this.modifiedDate,
+    required  this.isCreditNote,
     this.creditWith,
     this.creditAmount,
     this.creditCurrency,
