@@ -5,9 +5,14 @@ import 'package:floor/floor.dart';
 abstract class NotificationDao {
 
   @insert
-  Future<void> insertNotification(NotificationModelVO notificationModelVO);
+  Future<void> insertNotification(NotificationModelVO notification);
 
   @delete
-  Future<void> deleteNote(NotificationModelVO notificationModelVO);
+  Future<void> deleteNote(NotificationModelVO notification);
 
+  @Query('SELECT * FROM notification')
+  Future<List<NotificationModelVO>> findAllNoti();
+
+  @update
+  Future<int> updateNotification(NotificationModelVO notification);
 }
