@@ -21,8 +21,8 @@ class _PaymentViewState extends State<PaymentView> {
   var receipt = false;
   int changePageIndex = 0;
 
-   String currentInvoiceIndex ;
-   String currentTransactionIndex ;
+   late String currentInvoiceIndex ;
+   late String currentTransactionIndex ;
   final loginDataStorage = GetStorage();
 
   final InvoiceListController invoiceListController =
