@@ -29,7 +29,7 @@ class _SecondLoginVIewState extends State<SecondLoginVIew> {
 
   @override
   void initState() {
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
       if (langStorage.read(LANGUAGE) == 'မြန်မာ') {
         var locale = Locale('my', 'MM');
         Get.updateLocale(locale);
